refactor(data-encryption): split AES into block, key_expansion, and cipher translation units
  Extract Block_t/Word_t API into block.h/block.c, KeyExpansion_t API into
  key_expansion.h/key_expansion.c, and word-level helpers into the internal
  header src/word.h. Trim AES.h/AES.c to cipher-only (encryptBlock/decryptBlock).
  Mark SBox/GF256 lookup tables as static to prevent multiple-definition
  link errors when included from more than one translation unit.
